+++ NEWS.md
@@ -1,7 +1,7 @@
 # CHANGES IN GEOMORPH VERSION 3.1.3 (Patch Release)
 
 ### NEW FEATURES
-* New function added: `compare.CR 
+* New function added: `compare.CR` 
 
 ### OTHER CHANGES
 * Adjustment to getSurfPCs to allow for more robust surface sliding